[lib][dbconnector] Remove sql file and delegate SQL query to pg_connector
